analytics/requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
jupyter>=1.0.0
notebook>=6.4.0
openpyxl>=3.0.0
scikit-learn>=1.0.0
scipy>=1.7.0
psycopg2-binary>=2.9.9
SQLAlchemy>=2.0.0 